Add files via upload
The input calls class works similarly to a singly linked tree.  I found this to work much more efficiently for what I was doing. It follows methods almost identical to normal binary trees, aside from a few modifications.  The notification class acts as the nodes for the whole tree. This class has the single purpose of grouping all the nodes/notifications after they had been sorted and then outputting them similarly to a priority queue to the relevant departments.